+++ src/main/java/com/souldealers/crowdtracebackend/modules/identity/JwtFilter.java
@@ -7,8 +7,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.lang.NonNull;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -56,11 +54,7 @@ public class JwtFilter extends OncePerRequestFilter {
                 }
             }
         } catch (Exception e){
-            String errorMessage = "{\"error\": \"Forbidden\", \"message\": \"" + e.getMessage() + "\"}";
-            response.setStatus(HttpStatus.FORBIDDEN.value());
-            response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-            response.getWriter().write(errorMessage);
-            response.getWriter().flush();
+            SecurityContextHolder.clearContext();
         }
 
         filterChain.doFilter(request, response);