+++ src/main/java/com/souldealers/crowdtracebackend/modules/identity/SecurityConfig.java
@@ -8,6 +8,7 @@ import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
@@ -19,6 +20,7 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 
 @Configuration
 @RequiredArgsConstructor
+@EnableMethodSecurity
 public class SecurityConfig {
 
     private final JwtFilter jwtFilter;
@@ -34,7 +36,8 @@ public class SecurityConfig {
             "/api/v1/auth/verify-otp",
             "/api/v1/auth/resend-otp",
             "/api/v1/auth/request-password-reset",
-            "/api/v1/auth/reset-password"
+            "/api/v1/auth/reset-password",
+            "/api/public/**"
     };
 
     @Bean